refactor(components): replace checkbox tree mutation observer
Replace checkbox tree attribute observation with bubbled checkbox modification events.

Keep dynamic composition on slotchange and synchronize only from the outer recursive tree.

Add a modified-event-only mixin so checkbox state changes do not emit connection lifecycle events.

Compose the existing lifecycle mixin from the narrower helper without changing its event contract.
